+++ src/Heartbeat/Sensor/Status.php
@@ -16,9 +16,14 @@ class Status
     public readonly string $name;
 
     /**
-     * @var mixed The actual status
+     * @var bool The actual status
      */
-    public readonly mixed $status;
+    public readonly bool $status;
+
+    /**
+     * @var string The human-readable status message
+     */
+    public readonly string $message;
 
     /**
      * @var float The sensor check duration in seconds
@@ -44,23 +49,26 @@ class Status
      * Status construction
      *
      * @param string $name The name of the sensor
-     * @param mixed $status The actual status.
+     * @param bool $status The actual status.
      * @param float $duration How long it took to execute the check.
      * @param Chronos $lastExecuted The date/time when it was executed last.
      * @param Severity $severity The status severity.
+     * @param string $message The human-readable status message.
      */
     public function __construct(
         string $name,
-        mixed $status,
+        bool $status,
         float $duration,
         Chronos $lastExecuted,
         Severity $severity,
+        string $message,
     ) {
         $this->severity = $severity;
         $this->lastExecuted = $lastExecuted;
         $this->duration = $duration;
         $this->status = $status;
         $this->name = $name;
+        $this->message = $message;
     }
 
     /**